pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "garmin-buddy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "altair>=6.0.0",
    "dotenv>=0.9.9",
    "fitparse>=1.2.0",
    "garminconnect>=0.2.38",
    "google-genai>=1.60.0",
    "google-api-python-client>=2.169.0",
    "pandas>=2.3.3",
    "pyodbc>=5.3.0",
    "PyYAML>=6.0.2",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.53.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]


[tool.hatch.build.targets.wheel]
packages = ["src/garmin_buddy"]
